Updating source code and executables (hide column option and complete refactor).
